+++ README.md
@@ -33,7 +33,6 @@ Feel free to contribute by adding as many features as you want!
 * [commandSuggestorIgnoresSpaces](#commandsuggestorignoresspaces)
 * [commandTravel](#commandtravel)
 * [disableBobViewWhenHurt](#disablebobviewwhenhurt)
-* [disableFovChangeInWater](#disablefovchangeinwater)
 * [disableHotbarScrolling](#disablehotbarscrolling)
 * [disableJoinLeaveMessages](#disablejoinleavemessages)
 * [disableNarrator](#disablenarrator)
@@ -46,6 +45,8 @@ Feel free to contribute by adding as many features as you want!
 * [enableScriptOnJoin](#enablescriptonjoin)
 * [essentialClientMainMenu](#essentialclientmainmenu)
 * [highlightLavaSources](#highlightlavasource)
+* [waterFovMultipler](#waterFovMultipler)
+* [soulSpeedFovMultiplier](#soulSpeedFovMultipler)
 * [increaseSpectatorScrollSensitivity](#increasespectatorscrollsensitivity)
 * [increaseSpectatorScrollSpeed](#increasespectatorscrollspeed)
 * [missingTools](#missingtools)
@@ -159,11 +160,6 @@ Disables the camera bobbing when you get hurt
 * Type: `Boolean`
 * Default Value: `false`
 
-## disableFovChangeInWater
-This stops the FOV changing when you are submerged in water
-* Type: `Boolean`
-* Default Value: `false`
-
 ## disableHotbarScrolling
 This will prevent you from scrolling in your hotbar
 * Type: `Boolean`
@@ -235,6 +231,16 @@ This renders the Essential Client Menu on the main menu screen
 Highlights lava sources, credit to [plusls](https://github.com/plusls) for the original code for this in [their mod](https://github.com/plusls/oh-my-minecraft-client)
 * Type: `Boolean`
 * Default Value: `false`
+
+## waterFovMultipler
+Determines the percentage of Fov scaling when fully submerged in water
+* Type: `Integer`
+* Default Value: `0`
+
+## soulSpeedFovMultipler
+Determines the percentage of Fov scaling when walking on soil soul or soul sand
+* Type: `Integer`
+* Default Value: `0`
 
 ## increaseSpectatorScrollSensitivity
 Increases the sensitivity at which you can scroll to go faster in spectator
