+++ src/main/resources/essentialclient.mixins.json
@@ -23,8 +23,8 @@
     "core.MinecraftClientMixin",
     "core.TitleScreenMixin",
     "disableBobViewWhenHurt.GameRendererMixin",
-    "disableFovChangeInWater.GameRendererMixin",
-    "soulBlocksEffectOnFov.AbstractClientPlayerEntityMixin",
+    "waterFovMultiplier.GameRendererMixin",
+    "soulSpeedFovMultipler.AbstractClientPlayerEntityMixin",
     "disableNarrator.KeyboardMixin",
     "disableNightVisionFlash.GameRendererMixin",
     "disableRecipeNotifications.RecipeToastMixin",
